comic_service/requirements.txt
# MCP SDK
mcp>=0.9.0

# 数据验证
pydantic>=2.5.0

# HTTP 客户端（异步请求 Gemini API）
httpx>=0.26.0
aiohttp>=3.9.0

# 图片处理
Pillow>=10.2.0

# 图像相似度（一致性验证）
scikit-image>=0.22.0
opencv-python>=4.9.0

# 文本处理
jieba>=0.42.1

# 日志
loguru>=0.7.2

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 工具库
python-dotenv>=1.0.0
pyyaml>=6.0.1
